Install desired Jenkins plugins on startup
It turns out that, to install the Jenkins plugins we want, we just have
to download them to disk (under JENKINS_HOME/plugins) before attempting
to start Jenkins.  Except with a weird file extension change.

This is done as a ExecStartPre instead of a packer step because the
relevant directory doesn't exist yet (it's on a mount that only gets
attached during provisioning).

This makes first start take _much_ longer, but somehow systemd is okay
with that and doesn't kill us, so... meh.